Use `double` instead of `float`

+++ src/main/java/ua/mibal/booking/model/dto/search/ApartmentCardDto.java
@@ -31,7 +31,7 @@ public record ApartmentCardDto(
         String name,
         List<Photo> photos,
         ApartmentOptions options,
-        Float rating,
+        Double rating,
         List<Bed> beds,
         BigDecimal cost
 ) {
